backend/requirements.txt
flask==3.0.3
flask-cors==4.0.1
pandas==3.0.2
numpy==2.3.4
scikit-learn==1.7.0
joblib==1.4.2
Pillow>=10.0.0
torch>=2.3.0
torchvision>=0.18.0
